comment通知メールを非同期処理に変更

+++ app/helpers/header_helper.rb
@@ -17,6 +17,7 @@ module HeaderHelper
 
   def display_btn
     return tag.div(class:'header-right', id:'article-post-btn') { |tag| 'Post' } if current_page?(new_article_path)
+    return tag.div(class:'header-left') { |tag| link_to 'Follow', :back } if params[:controller] == 'profiles' && params[:action] == 'show' && params[:id].to_i != current_user.id
     return tag.div(class:'header-right') { |tag| link_to 'Logout', destroy_user_session_path , method: 'delete' } if (current_page?(articles_path) || current_page?(root_path)) && signed_in?
     return tag.div(class:'header-right') { |tag| link_to 'Login', user_session_path , method: 'post' } if (current_page?(articles_path) || current_page?(root_path)) && !signed_in?
     return tag.div(class:'header-right hidden')
